pyproject.toml
[project]
name = "nonprofit-data-pipeline"
version = "0.1.0"
description = "Nonprofit donor database integration and analytics pipeline"
requires-python = ">=3.14"
dependencies = [
    "altair>=4.2,<5",
    "faker>=40.1.2",
    "numpy>=2.4.1",
    "openai>=2.17.0",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.46",
    "standard-imghdr>=3.13.0",
    "streamlit>=1.19.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
